translate the hotswap test
